테스트케이스 CRUD

+++ backend/src/main/java/com/demo/sortgeek/domain/dto/response/CustomTestCaseResponseDTO.java
@@ -10,5 +10,5 @@ import lombok.*;
 @AllArgsConstructor
 public class CustomTestCaseResponseDTO {
     private Long testCaseId;  // 테스트 케이스 ID
-    private List<Integer> testInput;  // 테스트 입력 데이터
+    private List<Long> testInput;  // 테스트 입력 데이터
 }
